# ezkey-demo-app-acme/openapi-spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ezkey Admin API",
    "description": "Administration API for Ezkey - Open Source Cryptographic MFA Platform\n\nThis API enables administrative management of Ezkey's main entities:\n- **Integrations**: Applications or systems protected by MFA\n- **Enrollments**: Associations between users, devices and integrations\n- **Auth Attempts**: MFA authentication attempts\n- **Admin Management**: Administrator authentication,\n                        enrollment recovery, and admin operations\n\nEzkey is intentionally distinct from FIDO2/WebAuthn and follows its own\ncryptographic MFA model. The API uses REST conventions and DTOs for all\nrequests and responses.\n",
    "contact": {
      "name": "Ezkey Team",
      "url": "https://ezkey.org",
      "email": "contributors@ezkey.org"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:9080",
      "description": "Development server Admin API"
    },
    {
      "url": "https://admin-api.ezkey.org",
      "description": "Production server Admin API"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/integrations": {
      "get": {
        "tags": [
          "Integrations"
        ],
        "summary": "Search integrations",
        "description": "Retrieves integrations with optional filters and pagination for administration and compliance reporting. Supports dynamic sorting via ?sort=field,direction (e.g., ?sort=id,asc). Default sort is by creation date descending (newest first). Optional tenantId filter: GlobalAdmin only; TenantAdmin scope is always their tenant.",
        "operationId": "search",
        "parameters": [
          {
            "name": "integrationName",
            "in": "query",
            "description": "Filter by integration name (partial match, case-insensitive)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "active",
            "in": "query",
            "description": "Filter by active flag",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "createdAfter",
            "in": "query",
            "description": "Filter integrations created after this timestamp (ISO-8601)",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "createdBefore",
            "in": "query",
            "description": "Filter integrations created before this timestamp (ISO-8601)",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "tenantId",
            "in": "query",
            "description": "Filter by tenant ID. GlobalAdmin only; when provided limits results to that tenant. Ignored for TenantAdmin.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Zero-based page index (0..N)",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "minimum": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "The size of the page to be returned",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "minimum": 1
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
            "required": false,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "array",
              "default": [
                "createdAt,DESC"
              ],
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Search completed successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PagedModelIntegrationResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PagedModelIntegrationResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Integrations"
        ],
        "summary": "Create new integration",
        "description": "Creates a new integration with the provided data. Integration code must be unique per tenant.",
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IntegrationCreateRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Integration created successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationCreateResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid data (validation errors: invalid code format, blank code, etc.)",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Integration code already exists for this tenant",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/api-keys/{keyId}": {
      "get": {
        "tags": [
          "API Keys"
        ],
        "summary": "Get API key details",
        "description": "Returns details of a specific API key. Secret key is never included.",
        "operationId": "getApiKey",
        "parameters": [
          {
            "name": "keyId",
            "in": "path",
            "description": "API key ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "example": 42
          }
        ],
        "responses": {
          "200": {
            "description": "API key details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - admin token required",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "API key not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyResponseDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "API Keys"
        ],
        "summary": "Revoke API key",
        "description": "Immediately revokes an API key, making it unusable for authentication. The key is preserved for audit purposes.",
        "operationId": "revokeApiKey",
        "parameters": [
          {
            "name": "keyId",
            "in": "path",
            "description": "API key ID to revoke",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "example": 42
          },
          {
            "name": "reason",
            "in": "query",
            "description": "Audit justification for the revocation (min 10 characters)",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 500,
              "minLength": 10
            }
          }
        ],
        "responses": {
          "204": {
            "description": "API key revoked successfully"
          },
          "401": {
            "description": "Unauthorized - admin token required"
          },
          "404": {
            "description": "API key not found"
          },
          "429": {
            "description": "Too many revoke requests - rate limit exceeded"
          }
        }
      },
      "patch": {
        "tags": [
          "API Keys"
        ],
        "summary": "Partially update API key configuration",
        "description": "Updates API key config (ipWhitelist, description). Only active keys. Include version from GET for optimistic locking.",
        "operationId": "updateApiKey",
        "parameters": [
          {
            "name": "keyId",
            "in": "path",
            "description": "API key ID to update",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "example": 42
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiKeyUpdateRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "API key updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid data, key revoked, or ipWhitelist validation failed (invalid CIDR)",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Access denied",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "API key not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Optimistic lock conflict - resource was modified, re-fetch and retry",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Too many update requests - rate limit exceeded",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyResponseDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "IntegrationCreateRequestDto": {
        "type": "object",
        "description": "Integration creation data including code, name, and description",
        "properties": {
          "code": {
            "type": "string",
            "description": "Unique business identifier code for the integration",
            "example": "web-portal",
            "maxLength": 100,
            "minLength": 2,
            "pattern": "^[a-zA-Z0-9_-]+$"
          },
          "name": {
            "type": "string",
            "description": "Display name for the integration",
            "example": "Web Portal",
            "maxLength": 255,
            "minLength": 0
          },
          "description": {
            "type": "string",
            "description": "Optional description of the integration",
            "example": "Administration console",
            "maxLength": 500,
            "minLength": 0
          }
        },
        "required": [
          "code",
          "name"
        ]
      },
      "IntegrationCreateResponseDto": {
        "type": "object",
        "description": "Response DTO containing created integration details",
        "properties": {
          "code": {
            "type": "string",
            "description": "Unique business identifier code for the integration",
            "example": "web-portal"
          },
          "id": {
            "type": "integer",
            "format": "int32",
            "description": "Unique identifier of the newly created integration",
            "example": 42
          }
        }
      },
      "ErrorResponseDto": {
        "type": "object",
        "description": "Standardized error response for all API errors",
        "properties": {
          "code": {
            "type": "string",
            "description": "Error code identifier for categorizing the error type",
            "example": "RESOURCE_NOT_FOUND"
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message describing what went wrong",
            "example": "The requested resource was not found"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the error occurred",
            "example": "2025-10-15T14:30:00+00:00"
          },
          "path": {
            "type": "string",
            "description": "Request path that caused the error",
            "example": "/api/v1/auth-attempts/123"
          }
        },
        "required": [
          "code",
          "message",
          "timestamp"
        ]
      },
      "ProblemDetail": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string",
            "format": "uri"
          },
          "properties": {
            "type": "object",
            "additionalProperties": {}
          }
        }
      },
...
      "PageMetadata": {
        "type": "object",
        "properties": {
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "number": {
            "type": "integer",
            "format": "int64"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "PagedModelIntegrationResponseDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IntegrationResponseDto"
            }
          },
          "page": {
            "$ref": "#/components/schemas/PageMetadata"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT authentication token for admin API",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}